Distribute unit tests across separate files
