+++ tests/models/all_modeling_tests.js
@@ -1,2 +1,33 @@
+export { default as bert } from "./bert/test_modeling_bert.js";
+export { default as bloom } from "./bloom/test_modeling_bloom.js";
 export { default as clip } from "./clip/test_modeling_clip.js";
+export { default as codegen } from "./codegen/test_modeling_codegen.js";
+export { default as cohere } from "./cohere/test_modeling_cohere.js";
+export { default as florence2 } from "./florence2/test_modeling_florence2.js";
+export { default as gemma } from "./gemma/test_modeling_gemma.js";
+export { default as gemma2 } from "./gemma2/test_modeling_gemma2.js";
+export { default as gpt2 } from "./gpt2/test_modeling_gpt2.js";
+export { default as gpt_bigcode } from "./gpt_bigcode/test_modeling_gpt_bigcode.js";
+export { default as gpt_neo } from "./gpt_neo/test_modeling_gpt_neo.js";
+export { default as gpt_neox } from "./gpt_neox/test_modeling_gpt_neox.js";
+export { default as gptj } from "./gptj/test_modeling_gptj.js";
+export { default as granite } from "./granite/test_modeling_granite.js";
+export { default as idefics3 } from "./idefics3/test_modeling_idefics3.js";
+export { default as jais } from "./jais/test_modeling_jais.js";
+export { default as llama } from "./llama/test_modeling_llama.js";
+export { default as llava } from "./llava/test_modeling_llava.js";
+export { default as marian } from "./marian/test_modeling_marian.js";
+export { default as mistral } from "./mistral/test_modeling_mistral.js";
+export { default as mpt } from "./mpt/test_modeling_mpt.js";
+export { default as musicgen } from "./musicgen/test_modeling_musicgen.js";
+export { default as olmo } from "./olmo/test_modeling_olmo.js";
+export { default as olmo2 } from "./olmo2/test_modeling_olmo2.js";
+export { default as opt } from "./opt/test_modeling_opt.js";
+export { default as paligemma } from "./paligemma/test_modeling_paligemma.js";
+export { default as patchtsmixer } from "./patchtsmixer/test_modeling_patchtsmixer.js";
+export { default as patchtst } from "./patchtst/test_modeling_patchtst.js";
 export { default as pyannote } from "./pyannote/test_modeling_pyannote.js";
+export { default as qwen2_vl } from "./qwen2_vl/test_modeling_qwen2_vl.js";
+export { default as t5 } from "./t5/test_modeling_t5.js";
+export { default as vision_encoder_decoder } from "./vision_encoder_decoder/test_modeling_vision_encoder_decoder.js";
+export { default as whisper } from "./whisper/test_modeling_whisper.js";
